apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
PyYAML==6.0.2
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
SQLAlchemy==2.0.36
python-jose[cryptography]==3.3.0
ruamel.yaml==0.18.6
psycopg[binary]==3.2.3
